requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
distro==1.8.0
Flask==3.0.0
flask-ngrok==0.0.25
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.3.4
packaging==23.2
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pytesseract==0.3.10
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
